requirements.txt
tensorflow
numpy
pillow
pandas
tabulate
tqdm
wrapt
h5py
keras
